+++ CHANGELOG.md
@@ -1,11 +1,29 @@
+# v0.3.0
 
+* 支持action方法自定义传参
+* 完善readme文档
+
+---
+
+* support pass custom args in action function
+* improve the docs
 
 # v0.2.2
+
+* 优化缓存state和action对象
+* 添加更多的测试用例
+
+--- 
 
 * memorize the state & action
 * add more test case
 
 # v0.2.0
+
+* 支持自动监听嵌套的对象的值
+* 默认导出 hook 函数
+
+---
 
 * support obseve nested object
 * export state and action hooks as default